worker/pyproject.toml
[project]
name = "zeropath-worker"
version = "0.1.0"
description = "ZeroPath scanner worker – receives Pub/Sub push messages and runs scans"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115,<1",
    "uvicorn[standard]>=0.34,<1",
    "sqlalchemy>=2.0,<3",
    "psycopg2-binary>=2.9,<3",
    "pydantic>=2.0,<3",
    "pydantic-settings>=2.0,<3",
    "httpx>=0.27,<1",
    "openai>=1.60,<2",
    "python-dotenv>=1.0,<2",
    "gitpython>=3.1,<4",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.9,<1",
    "pytest>=8,<9",
    "pytest-asyncio>=0.24,<1",
]

[build-system]
requires = ["setuptools>=75"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
